# code cleanup/fixes improvements from work: - no longer export classes; instead inline functions where it makes sense (e.g. the scope timers) or export the member functions directly - fix icc11 warnings - add some comments
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@6531 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/sysdep/os/win/wdbg.cpp
@@ -24,10 +24,10 @@ int debug_IsPointerBogus(const void* p)
 	if(p < (void*)0x10000)
 		return true;
 #if ARCH_AMD64
-	if(p == (const void*)0xCCCCCCCCCCCCCCCCull)
+	if(p == (const void*)(uintptr_t)0xCCCCCCCCCCCCCCCCull)
 		return true;
 #elif ARCH_IA32
-	if(p == (const void*)0xCCCCCCCCul)
+	if(p == (const void*)(uintptr_t)0xCCCCCCCCul)
 		return true;
 #endif
